+++ internal/provider/resource_tfe_workspace_execution_mode_test.go
@@ -6,7 +6,6 @@ package provider
 import (
 	"errors"
 	"fmt"
-	"log"
 	"testing"
 
 	tfe "github.com/hashicorp/go-tfe"
@@ -23,6 +22,8 @@ func TestAccTFEWorkspaceExecutionMode_create_update(t *testing.T) {
 	org, orgCleanup := createBusinessOrganization(t, tfeClient)
 	t.Cleanup(orgCleanup)
 
+	attachPool := &tfe.Workspace{}
+
 	resource.Test(t, resource.TestCase{
 		PreCheck:  func() { testAccPreCheck(t) },
 		Providers: testAccProviders,
@@ -30,40 +31,47 @@ func TestAccTFEWorkspaceExecutionMode_create_update(t *testing.T) {
 			{
 				Config: testAccTFEWorkspaceExecutionMode_basic(org.Name),
 				Check: resource.ComposeTestCheckFunc(
-					testAccTFECheckWorkspaceAgentPoolAttached("tfe_workspace.workspace", "tfe_agent_pool.pool"),
+					testAccTFECheckWorkspaceAgentPoolAttached("tfe_workspace_execution_mode.attach", attachPool),
 					resource.TestCheckResourceAttr("tfe_agent_pool.pool", "organization_scoped", "false"),
+					resource.TestCheckResourceAttr("tfe_workspace_execution_mode.attach", "execution_mode", "agent"),
 				),
 			},
 			{
 				Config: testAccTFEWorkspaceExecutionMode_update(org.Name),
 				Check: resource.ComposeTestCheckFunc(
-					testAccTFECheckWorkspaceAgentPoolAttached("tfe_workspace.workspace", "tfe_agent_pool.pool"),
+					testAccTFECheckWorkspaceAgentPoolAttached("tfe_workspace_execution_mode.attach", attachPool),
 					resource.TestCheckResourceAttr("tfe_agent_pool.pool", "organization_scoped", "false"),
+					resource.TestCheckResourceAttr("tfe_workspace_execution_mode.attach", "execution_mode", "agent"),
 				),
 			},
 			{
 				Config: testAccTFEWorkspaceExecutionMode_destroy(org.Name),
 				Check: resource.ComposeTestCheckFunc(
-					testAccTFECheckWorkspaceAgentPoolNotDetached("tfe_workspace.workspace", "tfe_agent_pool.pool"),
+					testAccTFECheckWorkspaceAgentPoolNotDetached("tfe_workspace_execution_mode.attach", attachPool),
+					resource.TestCheckResourceAttr("tfe_workspace_execution_mode.attach", "execution_mode", "agent"),
 				),
 			},
 		},
 	})
 }
 
-func testAccTFECheckWorkspaceAgentPoolAttached(workspace string, pool string) resource.TestCheckFunc {
+func testAccTFECheckWorkspaceAgentPoolAttached(w string, attachPool *tfe.Workspace) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		config := testAccProvider.Meta().(ConfiguredClient)
 
 		// Read state file for workspace
-		ws, ok := s.RootModule().Resources[workspace]
+		ws, ok := s.RootModule().Resources[w]
 		if !ok {
-			return fmt.Errorf("Resource not found: %s", workspace)
+			return fmt.Errorf("Resource not found: %s", w)
 		}
 
 		// Resource ID equals the Workspace ID
 		if ws.Primary.ID == "" {
 			return fmt.Errorf("No instance ID is set")
+		}
+
+		if ws.Primary.Attributes["agent_pool_id"] == "" {
+			return fmt.Errorf("No Agent Pool ID is set")
 		}
 
 		workspace, err := config.Client.Workspaces.ReadByID(ctx, ws.Primary.ID)
@@ -71,48 +79,27 @@ func testAccTFECheckWorkspaceAgentPoolAttached(workspace string, pool string) re
 			return fmt.Errorf("error fetching workspace: %w", err)
 		}
 
-		fmt.Println("!!!!!!", workspace.AgentPoolID)
-
-		// Read state file for agent pool
-		ap, ok := s.RootModule().Resources[pool]
-		if !ok {
-			return fmt.Errorf("Resource not found: %s", pool)
+		if workspace.AgentPool.ID == "" {
+			return fmt.Errorf("error attaching agent pool inside attach fn %w", err)
 		}
 
-		// Resource ID equals the Agent Pool ID
-		if ap.Primary.ID == "" {
-			return fmt.Errorf("No instance ID is set")
-		}
-
-		if workspace.AgentPoolID != ap.Primary.ID {
-			return fmt.Errorf("error attaching agent pool %s: %w", ap.Primary.ID, err)
-		}
+		*attachPool = *workspace
 
 		return nil
 	}
 }
 
-func testAccTFECheckWorkspaceAgentPoolNotDetached(workspace string, pool string) resource.TestCheckFunc {
+func testAccTFECheckWorkspaceAgentPoolNotDetached(w string, detachPool *tfe.Workspace) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		config := testAccProvider.Meta().(ConfiguredClient)
 
-		ws, ok := s.RootModule().Resources[workspace]
+		ws, ok := s.RootModule().Resources[w]
 		if !ok {
-			return fmt.Errorf("Resource not found: %s", workspace)
-		}
-
-		ap, ok := s.RootModule().Resources[pool]
-		if !ok {
-			return fmt.Errorf("Resource not found: %s", pool)
+			return fmt.Errorf("Resource not found: %s", w)
 		}
 
 		// Resource ID equals the Workspace ID
 		if ws.Primary.ID == "" {
-			return fmt.Errorf("No instance ID is set")
-		}
-
-		// Resource ID equals the Agent Pool ID
-		if ap.Primary.ID == "" {
 			return fmt.Errorf("No instance ID is set")
 		}
 
@@ -122,14 +109,10 @@ func testAccTFECheckWorkspaceAgentPoolNotDetached(workspace string, pool string)
 		}
 
 		if workspace.AgentPoolID != "" {
-			return fmt.Errorf("error detaching agent pool %s: %w", ap.Primary.ID, err)
+			return fmt.Errorf("error detaching agent pool %w", err)
 		}
 
-		pool, err := config.Client.AgentPools.Read(ctx, ap.Primary.ID)
-		if err != nil && !errors.Is(err, tfe.ErrResourceNotFound) {
-			log.Printf("[DEBUB] THERE IS SOMETHING WRONG WITH THIS AGENT POOL %s", pool.ID)
-			return fmt.Errorf("error fetching agent pool: %w", err)
-		}
+		*detachPool = *workspace
 
 		return nil
 	}
@@ -154,11 +137,11 @@ resource "tfe_agent_pool_allowed_workspaces" "permit"{
 		tfe_workspace.workspace.id
 		]
 }
+
 resource "tfe_workspace_execution_mode" "attach"{
 	workspace_id = tfe_workspace.workspace.id
 	agent_pool_id = tfe_agent_pool_allowed_workspaces.permit.agent_pool_id
 	execution_mode = "agent"
-	depends_on = [tfe_agent_pool_allowed_workspaces.permit]
 }`, organization, organization)
 }
 
@@ -186,7 +169,6 @@ resource "tfe_workspace_execution_mode" "attach"{
 	workspace_id = tfe_workspace.workspace.id
 	agent_pool_id = tfe_agent_pool_allowed_workspaces.permit.agent_pool_id
 	execution_mode = "agent"
-	depends_on = [tfe_agent_pool_allowed_workspaces.permit]
 }`, organization, organization)
 }
 
@@ -214,6 +196,5 @@ func testAccTFEWorkspaceExecutionMode_destroy(organization string) string {
 		workspace_id = tfe_workspace.workspace.id
 		agent_pool_id = tfe_agent_pool_allowed_workspaces.permit.agent_pool_id
 		execution_mode = "agent"
-		depends_on = [tfe_agent_pool_allowed_workspaces.permit]
 	}`, organization, organization)
 }